feat: DegreeInternalService 추가 및 로직 이동
- 도메인에 가까운 로직 InternalService로 이동
- UserService.update() 테스트 수정
