+++ src/main/java/woowacrew/user/service/UserInternalService.java
@@ -3,20 +3,20 @@ package woowacrew.user.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacrew.user.domain.*;
-import woowacrew.user.domain.exception.DegreeBoundException;
 import woowacrew.user.service.exception.NotExistUserException;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
 public class UserInternalService {
-    private DegreeRepository degreeRepository;
     private UserRepository userRepository;
+    private DegreeInternalService degreeInternalService;
 
-    public UserInternalService(DegreeRepository degreeRepository, UserRepository userRepository) {
-        this.degreeRepository = degreeRepository;
+    public UserInternalService(UserRepository userRepository, DegreeInternalService degreeInternalService) {
         this.userRepository = userRepository;
+        this.degreeInternalService = degreeInternalService;
     }
 
     public User findByOauthId(String oauthId) {
@@ -41,8 +41,20 @@ public class UserInternalService {
         return userRepository.findByRoleAndNicknameNotNull(role);
     }
 
-    public Degree findDegreeByNumber(int numberOfDegree) {
-        return degreeRepository.findByNumber(numberOfDegree)
-                .orElseThrow(DegreeBoundException::new);
+    @Transactional
+    public void approveUserFor(Long userId, UserContext userContext, UserApproveDto userApproveDto) {
+        User user = findById(userId);
+        User admin = findById(userContext.getId());
+        Degree degree = degreeInternalService.findDegreeByNumber(userApproveDto.getDegreeNumber());
+
+        user.updateByAdmin(admin, userApproveDto.getRole(), degree);
+    }
+
+    @Transactional
+    public User update(Long userId, String nickname, LocalDate birthday) {
+        User user = findById(userId);
+        user.updateUserInfo(nickname, birthday);
+
+        return userRepository.save(user);
     }
 }
