+++ src/test/java/woowacrew/user/service/UserServiceTest.java
@@ -17,7 +17,7 @@ import java.time.format.DateTimeFormatter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -47,8 +47,9 @@ class UserServiceTest {
     @Test
     void 정상적으로_유저의_정보를_업데이트한다() {
         User user = new User("test", new Degree());
+        user.updateUserInfo("test nickname", LocalDate.of(1995, 6,8));
 
-        when(userInternalService.findById(anyLong())).thenReturn(user);
+        when(userInternalService.update(anyLong(), anyString(), any())).thenReturn(user);
 
         UserUpdateDto userUpdateDto = new UserUpdateDto("test nickname", "1995-06-08");
         UserResponseDto updateUser = userService.update(1L, userUpdateDto);
